/10.03 : Implement RenderComponent & Animation Object

+++ ProjectCA/Include/Core/Components/AIComponent.h
@@ -20,7 +20,7 @@ public:
 
 public:
 	//virtual bool Init(CActor* pOwner, const Types::tstring& strTag = TEXT("InputComponent")) override;
-	virtual void Update(float fDeltaTime) override;
+	virtual void Update(double fDeltaTime) override;
 	
 
 public:
