+++ ProjectCA/Include/Core/Components/InputComponent.h
@@ -14,7 +14,7 @@ public:
 	
 public:
 	virtual bool Init(CActor* pOwner, const Types::tstring& strTag = TEXT("InputComponent"));
-	virtual void Update(float fDeltaTime) = 0;
+	virtual void Update(double fDeltaTime) = 0;
