+++ ProjectCA/Include/Scene/Actor/CPlayer.h
@@ -15,7 +15,7 @@ public:
 	//virtual bool Init(const Types::ActorData&) override;
 	virtual bool PostInit(const Types::ActorData&, CGameScene*) override;
 	virtual bool Init() override;
-	virtual void Update(float fDeltaTime) override;
+	virtual void Update(double fDeltaTime) override;
 	virtual void Render(const HDC& hDC) override;
 	virtual void ActorBehavior() override;
